Separated out weight/BMI scripts into their own project folder. Moved feature store manager to become a monorepo util
