+++ pipeline/feature_store_implementations/create_feature_store.py
@@ -1,5 +1,5 @@
 from dotenv import load_dotenv
-from feature_store_manager import FeatureStoreManager
+from phmlondon.feature_store_manager import FeatureStoreManager
 
 from phmlondon.snow_utils import SnowflakeConnection
